chore: move the dummy api to development router
